requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LLM SDKs
anthropic>=0.40.0
openai>=1.50.0
litellm>=1.50.0
google-genai>=1.51.0

# Validation and serialization
pydantic>=2.6.0
pydantic-settings>=2.1.0

# HTTP client for tools
httpx>=0.27.0
aiohttp>=3.9.0

# Utilities
python-ulid>=2.2.0
python-dotenv>=1.0.0

# SSE support
sse-starlette>=2.0.0

# Web search (DuckDuckGo)
ddgs>=9.0.0

# HTML to markdown conversion
html2text>=2024.2.26
beautifulsoup4>=4.12.0

# Async utilities
anyio>=4.2.0

# Supabase integration
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
